+++ tests/unit/test_apm_config/test_apm_config_cnf_file.py
@@ -130,9 +130,7 @@ class TestSolarWindsApmConfigCnfFile:
         assert resulting_config.get("bufsize") == 2
         assert resulting_config.get("histogram_precision") == 2
         assert resulting_config.get("reporter_file_single") == 2
-        assert resulting_config.get("enable_sanitize_sql") == True
         assert resulting_config.get("proxy") == "http://foo-bar"
-        assert resulting_config.get("is_grpc_clean_hack_enabled") == True
 
         # update_transaction_filters was called
         mock_update_txn_filters.assert_called_once_with(fixture_cnf_dict)
@@ -177,9 +175,7 @@ class TestSolarWindsApmConfigCnfFile:
             "bufsize": "foo",
             "histogramPrecision": "foo",
             "reporterFileSingle": "foo",
-            "enableSanitizeSql": "foo",
             "proxy": "foo",
-            "isGrpcCleanHackEnabled": "foo",
         }
         mock_get_cnf_dict = mocker.patch(
             "solarwinds_apm.apm_config.SolarWindsApmConfig.get_cnf_dict"
@@ -209,9 +205,7 @@ class TestSolarWindsApmConfigCnfFile:
         assert resulting_config.get("bufsize") == -1
         assert resulting_config.get("histogram_precision") == -1
         assert resulting_config.get("reporter_file_single") == 0
-        assert resulting_config.get("enable_sanitize_sql") == True
         assert resulting_config.get("proxy") == ""
-        assert resulting_config.get("is_grpc_clean_hack_enabled") == False
         # Meanwhile these are pretty open
         assert resulting_config.get("collector") == "False"
         assert resulting_config.get("hostname_alias") == "False"
@@ -257,9 +251,7 @@ class TestSolarWindsApmConfigCnfFile:
             "SW_APM_BUFSIZE": "3",
             "SW_APM_HISTOGRAM_PRECISION": "3",
             "SW_APM_REPORTER_FILE_SINGLE": "3",
-            "SW_APM_ENABLE_SANITIZE_SQL": "false",
             "SW_APM_PROXY": "http://other-foo-bar",
-            "SW_APM_IS_GRPC_CLEAN_HACK_ENABLED": "false",
         })
         mock_update_txn_filters = mocker.patch(
             "solarwinds_apm.apm_config.SolarWindsApmConfig.update_transaction_filters"
@@ -298,9 +290,7 @@ class TestSolarWindsApmConfigCnfFile:
         assert resulting_config.get("bufsize") == 3
         assert resulting_config.get("histogram_precision") == 3
         assert resulting_config.get("reporter_file_single") == 3
-        assert resulting_config.get("enable_sanitize_sql") == False
         assert resulting_config.get("proxy") == "http://other-foo-bar"
-        assert resulting_config.get("is_grpc_clean_hack_enabled") == False
 
         # Restore old collector
         if old_collector:
@@ -339,9 +329,7 @@ class TestSolarWindsApmConfigCnfFile:
             "SW_APM_BUFSIZE": "other-foo-bar",
             "SW_APM_HISTOGRAM_PRECISION": "other-foo-bar",
             "SW_APM_REPORTER_FILE_SINGLE": "other-foo-bar",
-            "SW_APM_ENABLE_SANITIZE_SQL": "other-foo-bar",
             "SW_APM_PROXY": "other-foo-bar",
-            "SW_APM_IS_GRPC_CLEAN_HACK_ENABLED": "other-foo-bar",
         })
         mock_update_txn_filters = mocker.patch(
             "solarwinds_apm.apm_config.SolarWindsApmConfig.update_transaction_filters"
@@ -377,9 +365,7 @@ class TestSolarWindsApmConfigCnfFile:
         assert resulting_config.get("bufsize") == 2
         assert resulting_config.get("histogram_precision") == 2
         assert resulting_config.get("reporter_file_single") == 2
-        assert resulting_config.get("enable_sanitize_sql") == True
         assert resulting_config.get("proxy") == "http://foo-bar"
-        assert resulting_config.get("is_grpc_clean_hack_enabled") == True
 
         # These are still valid, so env_var > cnf_file
         assert resulting_config.get("collector") == "False"